requirements.txt
openai
faiss-cpu
python-dotenv
numpy
pypdf